Original change: https://android-review.googlesource.com/c/platform/system/libfmq/+/1431616

+++ tests/msgq_test_client.cpp
@@ -49,10 +49,10 @@ using android::hardware::details::waitForHwService;
 using android::hardware::kSynchronizedReadWrite;
 using android::hardware::kUnsynchronizedWrite;
 
-typedef android::AidlMessageQueue<uint16_t, kSynchronizedReadWrite> AidlMessageQueueSync;
-typedef android::AidlMessageQueue<uint16_t, kUnsynchronizedWrite> AidlMessageQueueUnsync;
-typedef android::hardware::MessageQueue<uint16_t, kSynchronizedReadWrite> MessageQueueSync;
-typedef android::hardware::MessageQueue<uint16_t, kUnsynchronizedWrite> MessageQueueUnsync;
+typedef android::AidlMessageQueue<int32_t, kSynchronizedReadWrite> AidlMessageQueueSync;
+typedef android::AidlMessageQueue<int32_t, kUnsynchronizedWrite> AidlMessageQueueUnsync;
+typedef android::hardware::MessageQueue<int32_t, kSynchronizedReadWrite> MessageQueueSync;
+typedef android::hardware::MessageQueue<int32_t, kUnsynchronizedWrite> MessageQueueUnsync;
 static const std::string kServiceName = "BnTestAidlMsgQ";
 static constexpr size_t kNumElementsInSyncQueue = 1024;
 
@@ -183,7 +183,7 @@ class ClientUnsyncTestBase<MessageQueueUnsync> : public ::testing::Test {
             return false;
         }
         service->getFmqUnsyncWrite(configureFmq /* configureFmq */,
-                                   [queue](bool ret, const MQDescriptorUnsync<uint16_t>& in) {
+                                   [queue](bool ret, const MQDescriptorUnsync<int32_t>& in) {
                                        ASSERT_TRUE(ret);
                                        *queue = new (std::nothrow) MessageQueueUnsync(in);
                                    });
@@ -266,7 +266,7 @@ class UnsynchronizedWriteClient : public ClientUnsyncTestBase<T> {
 /*
  * Utility function to verify data read from the fast message queue.
  */
-bool verifyData(uint16_t* data, size_t count) {
+bool verifyData(int32_t* data, size_t count) {
     for (size_t i = 0; i < count; i++) {
         if (data[i] != i) return false;
     }
@@ -276,7 +276,7 @@ bool verifyData(uint16_t* data, size_t count) {
 /*
  * Utility function to initialize data to be written to the FMQ
  */
-inline void initData(uint16_t* data, size_t count) {
+inline void initData(int32_t* data, size_t count) {
     for (size_t i = 0; i < count; i++) {
         data[i] = i;
     }
@@ -309,7 +309,7 @@ TYPED_TEST(UnsynchronizedWriteClientMultiProcess, MultipleReadersAfterOverflow) 
         ASSERT_TRUE(ret);
 
         // The following read should fail due to the overflow.
-        std::vector<uint16_t> readData(numMessagesMax);
+        std::vector<int32_t> readData(numMessagesMax);
         ASSERT_FALSE(queue->read(&readData[0], numMessagesMax));
 
         /*
@@ -344,7 +344,7 @@ TYPED_TEST(UnsynchronizedWriteClientMultiProcess, MultipleReadersAfterOverflow) 
         ASSERT_TRUE(queue->isValid());
 
         // This read should fail due to the write overflow.
-        std::vector<uint16_t> readData(dataLen);
+        std::vector<int32_t> readData(dataLen);
         ASSERT_FALSE(queue->read(&readData[0], dataLen));
 
         /*
@@ -379,7 +379,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWrite1) {
      */
     this->mService->requestBlockingRead(dataLen);
     {
-        std::array<uint16_t, dataLen> data = {0};
+        std::array<int32_t, dataLen> data = {0};
         ret = this->mQueue->writeBlocking(
                 data.data(), data.size(),
                 static_cast<uint32_t>(ITestMsgQ::EventFlagBits::FMQ_NOT_FULL),
@@ -388,7 +388,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWrite1) {
         ASSERT_TRUE(ret);
     }
     {
-        std::array<uint16_t, kNumElementsInSyncQueue> data = {0};
+        std::array<int32_t, kNumElementsInSyncQueue> data = {0};
         ret = this->mQueue->writeBlocking(
                 data.data(), data.size(),
                 static_cast<uint32_t>(ITestMsgQ::EventFlagBits::FMQ_NOT_FULL),
@@ -417,7 +417,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWrite2) {
     /* Cause a context switch to allow service to block */
     sched_yield();
     {
-        std::array<uint16_t, dataLen> data = {0};
+        std::array<int32_t, dataLen> data = {0};
         ret = this->mQueue->writeBlocking(data.data(), data.size());
         ASSERT_TRUE(ret);
     }
@@ -427,7 +427,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWrite2) {
      * kNumElementsInSyncQueue will succeed.
      */
     {
-        std::array<uint16_t, kNumElementsInSyncQueue> data = {0};
+        std::array<int32_t, kNumElementsInSyncQueue> data = {0};
         ret = this->mQueue->writeBlocking(data.data(), data.size(), 5000000000 /* timeOutNanos */);
         ASSERT_TRUE(ret);
     }
@@ -453,7 +453,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWriteRepeat1) {
      * Write 64 elements into the queue for the service to consume
      */
     {
-        std::array<uint16_t, dataLen> data = {0};
+        std::array<int32_t, dataLen> data = {0};
         for (size_t i = 0; i < writeCount; i++) {
             ret = this->mQueue->writeBlocking(
                     data.data(), data.size(),
@@ -468,7 +468,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWriteRepeat1) {
      * blocking write should be successful.
      */
     {
-        std::array<uint16_t, kNumElementsInSyncQueue> data = {0};
+        std::array<int32_t, kNumElementsInSyncQueue> data = {0};
         ret = this->mQueue->writeBlocking(
                 data.data(), data.size(),
                 static_cast<uint32_t>(ITestMsgQ::EventFlagBits::FMQ_NOT_FULL),
@@ -499,7 +499,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWriteRepeat2) {
      * Write 64 elements into the queue writeCount times
      */
     {
-        std::array<uint16_t, dataLen> data = {0};
+        std::array<int32_t, dataLen> data = {0};
         for (size_t i = 0; i < writeCount; i++) {
             ret = this->mQueue->writeBlocking(
                     data.data(), data.size(),
@@ -514,7 +514,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWriteRepeat2) {
      * blocking write should be successful.
      */
     {
-        std::array<uint16_t, kNumElementsInSyncQueue> data = {0};
+        std::array<int32_t, kNumElementsInSyncQueue> data = {0};
         ret = this->mQueue->writeBlocking(
                 data.data(), data.size(),
                 static_cast<uint32_t>(ITestMsgQ::EventFlagBits::FMQ_NOT_FULL),
@@ -544,7 +544,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWriteRepeat3) {
      * Write 64 elements into the queue writeCount times
      */
     {
-        std::array<uint16_t, dataLen> data = {0};
+        std::array<int32_t, dataLen> data = {0};
         for (size_t i = 0; i < writeCount; i++) {
             ret = this->mQueue->writeBlocking(
                     data.data(), data.size(),
@@ -559,7 +559,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWriteRepeat3) {
      * blocking write should be successful.
      */
     {
-        std::array<uint16_t, kNumElementsInSyncQueue> data = {0};
+        std::array<int32_t, kNumElementsInSyncQueue> data = {0};
         ret = this->mQueue->writeBlocking(
                 data.data(), data.size(),
                 static_cast<uint32_t>(ITestMsgQ::EventFlagBits::FMQ_NOT_FULL),
@@ -574,7 +574,7 @@ TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWriteRepeat3) {
  * attempts to write/read 0 messages and return true.
  */
 TYPED_TEST(SynchronizedReadWriteClient, BlockingReadWriteZeroMessages) {
-    uint16_t data = 0;
+    int32_t data = 0;
 
     /*
      * Trigger a blocking write for zero messages with no timeout.
@@ -602,7 +602,7 @@ TYPED_TEST(SynchronizedReadWriteClient, SmallInputReaderTest1) {
     ASSERT_LE(dataLen, kNumElementsInSyncQueue);
     bool ret = this->requestWriteFmqSync(dataLen);
     ASSERT_TRUE(ret);
-    uint16_t readData[dataLen] = {};
+    int32_t readData[dataLen] = {};
     ASSERT_TRUE(this->mQueue->read(readData, dataLen));
     ASSERT_TRUE(verifyData(readData, dataLen));
 }
@@ -646,7 +646,7 @@ TYPED_TEST(SynchronizedReadWriteClient, SmallInputWriterTest1) {
     const size_t dataLen = 16;
     ASSERT_LE(dataLen, kNumElementsInSyncQueue);
     size_t originalCount = this->mQueue->availableToWrite();
-    uint16_t data[dataLen];
+    int32_t data[dataLen];
     initData(data, dataLen);
     ASSERT_TRUE(this->mQueue->write(data, dataLen));
     bool ret = this->requestReadFmqSync(dataLen);
@@ -663,7 +663,7 @@ TYPED_TEST(SynchronizedReadWriteClient, SmallInputWriterTest2) {
     const size_t dataLen = 16;
     ASSERT_LE(dataLen, kNumElementsInSyncQueue);
     size_t originalCount = this->mQueue->availableToWrite();
-    uint16_t data[dataLen];
+    int32_t data[dataLen];
     initData(data, dataLen);
 
     typename TypeParam::MemTransaction tx;
@@ -673,8 +673,8 @@ TYPED_TEST(SynchronizedReadWriteClient, SmallInputWriterTest2) {
     auto second = tx.getSecondRegion();
 
     size_t firstRegionLength = first.getLength();
-    uint16_t* firstBaseAddress = first.getAddress();
-    uint16_t* secondBaseAddress = second.getAddress();
+    int32_t* firstBaseAddress = first.getAddress();
+    int32_t* secondBaseAddress = second.getAddress();
 
     for (size_t i = 0; i < dataLen; i++) {
         if (i < firstRegionLength) {
@@ -700,7 +700,7 @@ TYPED_TEST(SynchronizedReadWriteClient, ReadWhenEmpty) {
     ASSERT_EQ(0UL, this->mQueue->availableToRead());
     const size_t numMessages = 2;
     ASSERT_LE(numMessages, kNumElementsInSyncQueue);
-    uint16_t readData[numMessages];
+    int32_t readData[numMessages];
     ASSERT_FALSE(this->mQueue->read(readData, numMessages));
 }
 
@@ -714,7 +714,7 @@ TYPED_TEST(SynchronizedReadWriteClient, ReadWhenEmpty) {
  */
 
 TYPED_TEST(SynchronizedReadWriteClient, WriteWhenFull) {
-    std::array<uint16_t, kNumElementsInSyncQueue> data = {0};
+    std::array<int32_t, kNumElementsInSyncQueue> data = {0};
     initData(data.data(), data.size());
     ASSERT_TRUE(this->mQueue->write(data.data(), data.size()));
     ASSERT_EQ(0UL, this->mQueue->availableToWrite());
@@ -731,7 +731,7 @@ TYPED_TEST(SynchronizedReadWriteClient, WriteWhenFull) {
 TYPED_TEST(SynchronizedReadWriteClient, LargeInputTest1) {
     bool ret = this->requestWriteFmqSync(kNumElementsInSyncQueue);
     ASSERT_TRUE(ret);
-    std::vector<uint16_t> readData(kNumElementsInSyncQueue);
+    std::vector<int32_t> readData(kNumElementsInSyncQueue);
     ASSERT_TRUE(this->mQueue->read(&readData[0], kNumElementsInSyncQueue));
     ASSERT_TRUE(verifyData(&readData[0], kNumElementsInSyncQueue));
 }
@@ -747,7 +747,7 @@ TYPED_TEST(SynchronizedReadWriteClient, LargeInputTest2) {
     ASSERT_GT(numMessages, kNumElementsInSyncQueue);
     bool ret = this->requestWriteFmqSync(numMessages);
     ASSERT_FALSE(ret);
-    uint16_t readData;
+    int32_t readData;
     ASSERT_EQ(0UL, this->mQueue->availableToRead());
     ASSERT_FALSE(this->mQueue->read(&readData, 1));
 }
@@ -761,7 +761,7 @@ TYPED_TEST(SynchronizedReadWriteClient, LargeInputTest2) {
  */
 
 TYPED_TEST(SynchronizedReadWriteClient, LargeInputTest3) {
-    std::array<uint16_t, kNumElementsInSyncQueue> data = {0};
+    std::array<int32_t, kNumElementsInSyncQueue> data = {0};
     initData(data.data(), data.size());
     ASSERT_TRUE(this->mQueue->write(data.data(), data.size()));
     ASSERT_EQ(0UL, this->mQueue->availableToWrite());
@@ -785,7 +785,7 @@ TYPED_TEST(SynchronizedReadWriteClient, MultipleRead) {
     ASSERT_EQ(expectedCount, availableToRead);
     bool ret = this->requestWriteFmqSync(numMessages);
     ASSERT_TRUE(ret);
-    uint16_t readData[numMessages] = {};
+    int32_t readData[numMessages] = {};
     for (size_t i = 0; i < chunkNum; i++) {
         ASSERT_TRUE(this->mQueue->read(readData + i * chunkSize, chunkSize));
     }
@@ -801,7 +801,7 @@ TYPED_TEST(SynchronizedReadWriteClient, MultipleWrite) {
     const size_t chunkNum = 5;
     const size_t numMessages = chunkSize * chunkNum;
     ASSERT_LE(numMessages, kNumElementsInSyncQueue);
-    uint16_t data[numMessages];
+    int32_t data[numMessages];
     initData(&data[0], numMessages);
 
     for (size_t i = 0; i < chunkNum; i++) {
@@ -819,7 +819,7 @@ TYPED_TEST(SynchronizedReadWriteClient, MultipleWrite) {
  */
 TYPED_TEST(SynchronizedReadWriteClient, ReadWriteWrapAround) {
     size_t numMessages = kNumElementsInSyncQueue / 2;
-    std::array<uint16_t, kNumElementsInSyncQueue> data = {0};
+    std::array<int32_t, kNumElementsInSyncQueue> data = {0};
     initData(data.data(), data.size());
     ASSERT_TRUE(this->mQueue->write(&data[0], numMessages));
     bool ret = this->requestReadFmqSync(numMessages);
@@ -839,7 +839,7 @@ TYPED_TEST(SynchronizedReadWriteClient, ReadWriteWrapAround) {
  */
 TYPED_TEST(SynchronizedReadWriteClient, ReadWriteWrapAround2) {
     size_t numMessages = kNumElementsInSyncQueue / 2;
-    std::array<uint16_t, kNumElementsInSyncQueue> data = {0};
+    std::array<int32_t, kNumElementsInSyncQueue> data = {0};
     initData(data.data(), data.size());
     ASSERT_TRUE(this->mQueue->write(&data[0], numMessages));
     auto ret = this->requestReadFmqSync(numMessages);
@@ -856,7 +856,7 @@ TYPED_TEST(SynchronizedReadWriteClient, ReadWriteWrapAround2) {
     ASSERT_EQ(tx.getFirstRegion().getLength() + tx.getSecondRegion().getLength(), data.size());
 
     for (size_t i = 0; i < data.size(); i++) {
-        uint16_t* ptr = tx.getSlot(i);
+        int32_t* ptr = tx.getSlot(i);
         *ptr = data[i];
     }
 
@@ -875,7 +875,7 @@ TYPED_TEST(UnsynchronizedWriteClient, SmallInputReaderTest1) {
     ASSERT_LE(dataLen, this->mNumMessagesMax);
     bool ret = this->requestWriteFmqUnsync(dataLen, this->mService);
     ASSERT_TRUE(ret);
-    uint16_t readData[dataLen] = {};
+    int32_t readData[dataLen] = {};
     ASSERT_TRUE(this->mQueue->read(readData, dataLen));
     ASSERT_TRUE(verifyData(readData, dataLen));
 }
@@ -887,7 +887,7 @@ TYPED_TEST(UnsynchronizedWriteClient, SmallInputReaderTest1) {
 TYPED_TEST(UnsynchronizedWriteClient, SmallInputWriterTest1) {
     const size_t dataLen = 16;
     ASSERT_LE(dataLen, this->mNumMessagesMax);
-    uint16_t data[dataLen];
+    int32_t data[dataLen];
     initData(data, dataLen);
     ASSERT_TRUE(this->mQueue->write(data, dataLen));
     bool ret = this->requestReadFmqUnsync(dataLen, this->mService);
@@ -901,7 +901,7 @@ TYPED_TEST(UnsynchronizedWriteClient, ReadWhenEmpty) {
     ASSERT_EQ(0UL, this->mQueue->availableToRead());
     const size_t numMessages = 2;
     ASSERT_LE(numMessages, this->mNumMessagesMax);
-    uint16_t readData[numMessages];
+    int32_t readData[numMessages];
     ASSERT_FALSE(this->mQueue->read(readData, numMessages));
 }
 
@@ -915,7 +915,7 @@ TYPED_TEST(UnsynchronizedWriteClient, ReadWhenEmpty) {
  */
 
 TYPED_TEST(UnsynchronizedWriteClient, WriteWhenFull) {
-    std::vector<uint16_t> data(this->mNumMessagesMax);
+    std::vector<int32_t> data(this->mNumMessagesMax);
     initData(&data[0], this->mNumMessagesMax);
     ASSERT_TRUE(this->mQueue->write(&data[0], this->mNumMessagesMax));
     ASSERT_EQ(0UL, this->mQueue->availableToWrite());
@@ -932,7 +932,7 @@ TYPED_TEST(UnsynchronizedWriteClient, WriteWhenFull) {
 TYPED_TEST(UnsynchronizedWriteClient, LargeInputTest1) {
     bool ret = this->requestWriteFmqUnsync(this->mNumMessagesMax, this->mService);
     ASSERT_TRUE(ret);
-    std::vector<uint16_t> data(this->mNumMessagesMax);
+    std::vector<int32_t> data(this->mNumMessagesMax);
     ASSERT_TRUE(this->mQueue->read(&data[0], this->mNumMessagesMax));
     ASSERT_TRUE(verifyData(&data[0], this->mNumMessagesMax));
 }
@@ -947,7 +947,7 @@ TYPED_TEST(UnsynchronizedWriteClient, LargeInputTest2) {
     const size_t numMessages = this->mNumMessagesMax + 1;
     bool ret = this->requestWriteFmqUnsync(numMessages, this->mService);
     ASSERT_FALSE(ret);
-    uint16_t readData;
+    int32_t readData;
     ASSERT_EQ(0UL, this->mQueue->availableToRead());
     ASSERT_FALSE(this->mQueue->read(&readData, 1));
 }
@@ -962,7 +962,7 @@ TYPED_TEST(UnsynchronizedWriteClient, LargeInputTest2) {
  * to check if the reader process can recover from the error condition.
  */
 TYPED_TEST(UnsynchronizedWriteClient, LargeInputTest3) {
-    std::vector<uint16_t> data(this->mNumMessagesMax);
+    std::vector<int32_t> data(this->mNumMessagesMax);
     initData(&data[0], this->mNumMessagesMax);
     ASSERT_TRUE(this->mQueue->write(&data[0], this->mNumMessagesMax));
     ASSERT_EQ(0UL, this->mQueue->availableToWrite());
@@ -990,7 +990,7 @@ TYPED_TEST(UnsynchronizedWriteClient, MultipleRead) {
     ASSERT_EQ(expectedCount, availableToRead);
     bool ret = this->requestWriteFmqUnsync(numMessages, this->mService);
     ASSERT_TRUE(ret);
-    uint16_t readData[numMessages] = {};
+    int32_t readData[numMessages] = {};
     for (size_t i = 0; i < chunkNum; i++) {
         ASSERT_TRUE(this->mQueue->read(readData + i * chunkSize, chunkSize));
     }
@@ -1006,7 +1006,7 @@ TYPED_TEST(UnsynchronizedWriteClient, MultipleWrite) {
     const size_t chunkNum = 5;
     const size_t numMessages = chunkSize * chunkNum;
     ASSERT_LE(numMessages, this->mNumMessagesMax);
-    uint16_t data[numMessages];
+    int32_t data[numMessages];
     initData(data, numMessages);
     for (size_t i = 0; i < chunkNum; i++) {
         ASSERT_TRUE(this->mQueue->write(data + i * chunkSize, chunkSize));
@@ -1023,7 +1023,7 @@ TYPED_TEST(UnsynchronizedWriteClient, MultipleWrite) {
  */
 TYPED_TEST(UnsynchronizedWriteClient, ReadWriteWrapAround) {
     size_t numMessages = this->mNumMessagesMax / 2;
-    std::vector<uint16_t> data(this->mNumMessagesMax);
+    std::vector<int32_t> data(this->mNumMessagesMax);
     initData(&data[0], this->mNumMessagesMax);
     ASSERT_TRUE(this->mQueue->write(&data[0], numMessages));
     bool ret = this->requestReadFmqUnsync(numMessages, this->mService);
@@ -1052,14 +1052,14 @@ TYPED_TEST(UnsynchronizedWriteClient, SmallInputMultipleReaderTest) {
     pid_t pid;
     if ((pid = fork()) == 0) {
         /* child process */
-        uint16_t readData[dataLen] = {};
+        int32_t readData[dataLen] = {};
         ASSERT_TRUE(mQueue2->read(readData, dataLen));
         ASSERT_TRUE(verifyData(readData, dataLen));
         exit(0);
     } else {
         ASSERT_GT(pid,
                   0 /* parent should see PID greater than 0 for a good fork */);
-        uint16_t readData[dataLen] = {};
+        int32_t readData[dataLen] = {};
         ASSERT_TRUE(this->mQueue->read(readData, dataLen));
         ASSERT_TRUE(verifyData(readData, dataLen));
     }
@@ -1082,12 +1082,12 @@ TYPED_TEST(UnsynchronizedWriteClient, OverflowNotificationTest) {
     pid_t pid;
     if ((pid = fork()) == 0) {
         /* child process */
-        std::vector<uint16_t> readData(this->mNumMessagesMax);
+        std::vector<int32_t> readData(this->mNumMessagesMax);
         ASSERT_FALSE(mQueue2->read(&readData[0], this->mNumMessagesMax));
         exit(0);
     } else {
         ASSERT_GT(pid, 0/* parent should see PID greater than 0 for a good fork */);
-        std::vector<uint16_t> readData(this->mNumMessagesMax);
+        std::vector<int32_t> readData(this->mNumMessagesMax);
         ASSERT_FALSE(this->mQueue->read(&readData[0], this->mNumMessagesMax));
     }
 }
